V1.2.0

+++ Core API/Team7/Controllers/SaleItemController.cs
@@ -1,15 +1,13 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Team7.Models.Repository;
 using Team7.Models;
-using Microsoft.EntityFrameworkCore;
-using Team7.Context;
 using System.IO;
 using System.Net.Http.Headers;
+using Team7.Services;
 
 namespace Team7.Controllers
 {
@@ -32,10 +30,15 @@ namespace Team7.Controllers
         {
             try
             {
-                
+
                 SaleItemRepo.Add(saleItem);
-                await SaleItemRepo.SaveChangesAsync();
-                return Ok(saleItem);
+                if (await SaleItemRepo.SaveChangesAsync())
+                {
+                    return Ok();
+                } else
+                {
+                    return StatusCode(StatusCodes.Status503ServiceUnavailable, "Unable to add value in the database. Contact support.");
+                }
             }
             catch (Exception err)
             {
@@ -54,6 +57,7 @@ namespace Team7.Controllers
                 var file = formCollection.Files.First();
                 var folderName = Path.Combine("Resources", "Images", "saleItemImages");
                 var pathToSave = Path.Combine(Directory.GetCurrentDirectory(), folderName);
+
                 if (file.Length > 0)
                 {
                     var fileName = ContentDispositionHeaderValue.Parse(file.ContentDisposition).FileName.Trim('"');
@@ -70,50 +74,51 @@ namespace Team7.Controllers
                     return BadRequest();
                 }
             }
-            catch (Exception ex)
+            catch (Exception err)
             {
-                return StatusCode(500, $"Internal server error: {ex}");
+                return StatusCode(StatusCodes.Status500InternalServerError, err.Message);
             }
         }
 
         [HttpDelete]
         [Route("deletephoto")]
-        public async Task<IActionResult> Delphoto(string id)
+        public async Task<IActionResult> Delphoto(string name)
         {
-            System.IO.File.Delete(Path.Combine("Resources", "Images", "saleItemImages", id));
+            System.IO.File.Delete(Path.Combine("Resources", "Images", "saleItemImages", name));
             return Ok();
-        }
-
-        static void createFile(byte[] data)
-        {
-            using var stream = System.IO.File.Create(PATH);
-            stream.Write(data, 0, data.Length);
         }
 
         // PUT api/SaleItem/update/5
         [HttpPut]
         [Route("update")]
-        public async Task<IActionResult> PutSaleItem(SaleItem saleItem)
+        public async Task<IActionResult> PutSaleItem(int id, SaleItem saleItem)
         {
-            var toUpdate = await SaleItemRepo.GetSaleItemIdAsync(saleItem.SaleItemID);
+            var toUpdate = await SaleItemRepo._GetSaleItemIdAsync(id);
 
             if (toUpdate == null)
             {
-                return NotFound("Could not find existing sale item with id:");
+                return NotFound("Could not find existing sale item with ID - " + saleItem.SaleItemID);
             }
 
             try
             {
+
                 toUpdate.Name = saleItem.Name;
                 toUpdate.Photo = saleItem.Photo;
                 toUpdate.Description = saleItem.Description;
                 toUpdate.Price = saleItem.Price;
                 toUpdate.Quotable = saleItem.Quotable;
-                toUpdate.Quantity = saleItem.Quantity;
+                toUpdate.Quantity = toUpdate.Quantity;
+                //toUpdate.QuantityOnHand = saleItem.QuantityOnHand;
 
-                //VenueRepo.Update<Venue>(tempVenue);
-                await SaleItemRepo.SaveChangesAsync();
-                return Ok("Successfully updated");
+                SaleItemRepo.Update<SaleItem>(toUpdate);
+                if (await SaleItemRepo.SaveChangesAsync())
+                {
+                    return Ok();
+                } else
+                {
+                    return StatusCode(StatusCodes.Status503ServiceUnavailable, "Unable to update value in the database. Contact support.");
+                }
             }
             catch (Exception err)
             {
@@ -122,13 +127,12 @@ namespace Team7.Controllers
 
         }
 
-
         // DELETE api/SaleItem/delete/5
         [HttpDelete]
         [Route("delete")]
         public async Task<IActionResult> DeleteSaleItem(int id)
         {
-            var tempSaleItem = await SaleItemRepo.GetSaleItemIdAsync(id);
+            var tempSaleItem = await SaleItemRepo._GetSaleItemIdAsync(id);
             if (tempSaleItem == null)
             {
                 return NotFound();
@@ -136,12 +140,16 @@ namespace Team7.Controllers
             try
             {
                 SaleItemRepo.Delete<SaleItem>(tempSaleItem);
-                await SaleItemRepo.SaveChangesAsync();
-
-                var fileToDelete = tempSaleItem.Photo;
-                System.IO.File.Delete(Path.Combine("Resources", "Images", "saleItemImages", fileToDelete));
-
-                return Ok();
+                if (await SaleItemRepo.SaveChangesAsync())
+                {
+                    var fileToDelete = tempSaleItem.Photo;
+                    System.IO.File.Delete(Path.Combine("Resources", "Images", "saleItemImages", fileToDelete));
+                    return Ok();
+                }
+                else
+                {
+                    return StatusCode(StatusCodes.Status503ServiceUnavailable, "Unable to delete value in the database. Contact support.");
+                }
             }
             catch (Exception err)
             {
@@ -173,12 +181,18 @@ namespace Team7.Controllers
         // GET: api/SaleItem/getMatch/{input}
         [HttpGet]
         [Route("getMatch")]
-        public async Task<IActionResult> GetMatchingSaleItems(string input)
+        public async Task<IActionResult> GetMatchingSaleItems(string name, string description)
         {
             try
             {
-                var saleItem = await SaleItemRepo.GetSaleItemsAsync(input);
-                return Ok(saleItem);
+                var saleItem = await SaleItemRepo.GetSaleItemsAsync(name, description);
+                if (saleItem == null)
+                {
+                    return Ok(0);
+                } else
+                {
+                    return Ok(saleItem);
+                }
             }
             catch (Exception err)
             {
@@ -187,11 +201,4 @@ namespace Team7.Controllers
 
         }
 
-        [HttpGet]
-        [Route("exists")]
-        public async Task<SaleItem> SaleItemExists(int id)
-        {
-            return await SaleItemRepo.GetSaleItemIdAsync(id);
-        }
     }
-}
